requirements.txt
azure-storage-blob==12.19.0
langchain
openai
python-dotenv
streamlit
streamlit_chat
